backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
PyMySQL==1.4.6
